# rest-orchest-client/build/generated/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: teamtyrell@telekom-digital.com
    name: DOT Team Tyrell
  description: REST interface for interacting with Orchest Engine
  license:
    name: Proprietary license
  title: OrchesT
  version: 1.0.0
  x-api-category: other
servers:
- url: http://localhost:6200/orchest/
- url: https://api-orchest.your-domain.example.com/orchest/
- url: https://api-orchest.your-domain.example.com/orchest/
...
paths:
...
  /processDefinitionsSensitiveVariables:
    delete:
      operationId: deleteProcessSensitiveVariables
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessSensitiveVariables"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOString"
          description: OK
      tags:
      - process-definition-sensitive-variables
      x-content-type: application/json
      x-accepts:
      - '*/*'
    get:
      operationId: getAll
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessSensitiveVariables"
          description: OK
      tags:
      - process-definition-sensitive-variables
      x-accepts:
      - '*/*'
    patch:
      operationId: updateProcessSensitiveVariables
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessSensitiveVariables"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOProcessSensitiveVariables"
          description: OK
      tags:
      - process-definition-sensitive-variables
      x-content-type: application/json
      x-accepts:
      - '*/*'
    post:
      operationId: addProcessSensitiveVariables
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessSensitiveVariables"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOProcessSensitiveVariables"
          description: OK
      tags:
      - process-definition-sensitive-variables
      x-content-type: application/json
      x-accepts:
      - '*/*'
...
  /variables/{processInstanceId}:
    get:
      operationId: getVariables
      parameters:
      - explode: false
        in: path
        name: processInstanceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOVariablesResponse"
          description: OK
      tags:
      - variables-api
      x-accepts:
      - '*/*'
...
  /userTasks/process/{processInstanceId}:
    get:
      operationId: getTasksForProcessInstance
      parameters:
      - explode: false
        in: path
        name: processInstanceId
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          default: -createdAt
          type: string
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserTaskDTO"
          description: OK
      summary: List tasks for process instance
      tags:
      - User Tasks
      x-accepts:
      - '*/*'
...
  /processDefinitionsSensitiveVariables/{processDefinitionId}:
    get:
      operationId: getProcessSensitiveVariables
      parameters:
      - explode: false
        in: path
        name: processDefinitionId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOListProcessSensitiveVariables"
          description: OK
      tags:
      - process-definition-sensitive-variables
      x-accepts:
      - '*/*'
...
  /processDefinitions/{processDefinitionId}/{version}:
    delete:
      operationId: deleteProcessDefinition
      parameters:
      - explode: false
        in: path
        name: processDefinitionId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: version
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOVoid"
          description: OK
      tags:
      - process-definition-api
      x-accepts:
      - '*/*'
    get:
      operationId: getProcessDefinition_1
      parameters:
      - explode: false
        in: path
        name: processDefinitionId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: version
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseDTOResourceDefinitionDTO"
          description: OK
      tags:
      - process-definition-api
      x-accepts:
      - '*/*'
...
components:
  schemas:
...
    Meta:
      example:
        code: 0
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
    ResponseDTOVariablesResponse:
      example:
        data:
          variables:
            key: ""
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          $ref: "#/components/schemas/VariablesResponse"
    VariablesResponse:
      example:
        variables:
          key: ""
      properties:
        variables:
          additionalProperties: {}
...
    UserTaskDTO:
      example:
        claimedAt: claimedAt
        processInstanceId: processInstanceId
        processDefinitionId: processDefinitionId
        variables:
          key: ""
        completedAt: completedAt
        formKey: formKey
        dueDate: dueDate
        followUpDate: followUpDate
        activityId: activityId
        createdAt: createdAt
        candidateUsers:
        - candidateUsers
        - candidateUsers
        candidateGroups:
        - candidateGroups
        - candidateGroups
        taskName: taskName
        state: state
        assignee: assignee
        claimedBy: claimedBy
        taskId: taskId
      properties:
        taskId:
          type: string
        processInstanceId:
          type: string
        processDefinitionId:
          type: string
        activityId:
          type: string
        taskName:
          type: string
        state:
          type: string
        assignee:
          type: string
        candidateUsers:
          items:
            type: string
          type: array
        candidateGroups:
          items:
            type: string
          type: array
        claimedBy:
          type: string
        dueDate:
          type: string
        followUpDate:
          type: string
        formKey:
          type: string
        variables:
          additionalProperties: {}
        createdAt:
          type: string
        claimedAt:
          type: string
        completedAt:
          type: string
...
    ProcessSensitiveVariables:
      example:
        processDefinitionId: processDefinitionId
        createdAt: 2000-01-23T04:56:07.000+00:00
        variables:
        - name: name
          type: HIDE
          value: value
        - name: name
          type: HIDE
          value: value
        id: id
        enabled: true
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          type: string
        processDefinitionId:
          type: string
        variables:
          items:
            $ref: "#/components/schemas/SensitiveVariables"
          type: array
        enabled:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
    SensitiveVariables:
      example:
        name: name
        type: HIDE
        value: value
      properties:
        name:
          type: string
        value:
          type: string
        type:
          enum:
          - HIDE
          - SHOW
          type: string
    ResponseDTOProcessSensitiveVariables:
      example:
        data:
          processDefinitionId: processDefinitionId
          createdAt: 2000-01-23T04:56:07.000+00:00
          variables:
          - name: name
            type: HIDE
            value: value
          - name: name
            type: HIDE
            value: value
          id: id
          enabled: true
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          $ref: "#/components/schemas/ProcessSensitiveVariables"
...
    ResponseDTOString:
      example:
        data: data
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          type: string
...
    PageUserTaskDTO:
      example:
        number: 5
        last: true
        numberOfElements: 1
        size: 5
        totalPages: 0
        pageable:
          paged: true
          pageNumber: 1
          offset: 5
          pageSize: 5
          unpaged: true
          sort:
            unsorted: true
            sorted: true
            empty: true
        sort:
          unsorted: true
          sorted: true
          empty: true
        first: true
        content:
        - claimedAt: claimedAt
          processInstanceId: processInstanceId
          processDefinitionId: processDefinitionId
          variables:
            key: ""
          completedAt: completedAt
          formKey: formKey
          dueDate: dueDate
          followUpDate: followUpDate
          activityId: activityId
          createdAt: createdAt
          candidateUsers:
          - candidateUsers
          - candidateUsers
          candidateGroups:
          - candidateGroups
          - candidateGroups
          taskName: taskName
          state: state
          assignee: assignee
          claimedBy: claimedBy
          taskId: taskId
        - claimedAt: claimedAt
          processInstanceId: processInstanceId
          processDefinitionId: processDefinitionId
          variables:
            key: ""
          completedAt: completedAt
          formKey: formKey
          dueDate: dueDate
          followUpDate: followUpDate
          activityId: activityId
          createdAt: createdAt
          candidateUsers:
          - candidateUsers
          - candidateUsers
          candidateGroups:
          - candidateGroups
          - candidateGroups
          taskName: taskName
          state: state
          assignee: assignee
          claimedBy: claimedBy
          taskId: taskId
        totalElements: 6
        empty: true
      properties:
        totalPages:
          format: int32
          type: integer
        totalElements:
          format: int64
          type: integer
        pageable:
          $ref: "#/components/schemas/PageableObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        content:
          items:
            $ref: "#/components/schemas/UserTaskDTO"
          type: array
        number:
          format: int32
          type: integer
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    PageableObject:
      example:
        paged: true
        pageNumber: 1
        offset: 5
        pageSize: 5
        unpaged: true
        sort:
          unsorted: true
          sorted: true
          empty: true
      properties:
        paged:
          type: boolean
        pageNumber:
          format: int32
          type: integer
        pageSize:
          format: int32
          type: integer
        unpaged:
          type: boolean
        offset:
          format: int64
          type: integer
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      example:
        unsorted: true
        sorted: true
        empty: true
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    ResourceDefinitionDTO:
      example:
        resourceMetadata:
          key: ""
        createdAt: 2000-01-23T04:56:07.000+00:00
        resourceXML: resourceXML
        version: 5
        definitionId: definitionId
      properties:
        definitionId:
          type: string
        resourceXML:
          type: string
        version:
          format: int32
          type: integer
        createdAt:
          format: date-time
          type: string
        resourceMetadata:
          additionalProperties: {}
    PageProcessSensitiveVariables:
      example:
        number: 9
        last: true
        numberOfElements: 2
        size: 7
        totalPages: 0
        pageable:
          paged: true
          pageNumber: 1
          offset: 5
          pageSize: 5
          unpaged: true
          sort:
            unsorted: true
            sorted: true
            empty: true
        sort:
          unsorted: true
          sorted: true
          empty: true
        first: true
        content:
        - processDefinitionId: processDefinitionId
          createdAt: 2000-01-23T04:56:07.000+00:00
          variables:
          - name: name
            type: HIDE
            value: value
          - name: name
            type: HIDE
            value: value
          id: id
          enabled: true
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - processDefinitionId: processDefinitionId
          createdAt: 2000-01-23T04:56:07.000+00:00
          variables:
          - name: name
            type: HIDE
            value: value
          - name: name
            type: HIDE
            value: value
          id: id
          enabled: true
          updatedAt: 2000-01-23T04:56:07.000+00:00
        totalElements: 6
        empty: true
      properties:
        totalPages:
          format: int32
          type: integer
        totalElements:
          format: int64
          type: integer
        pageable:
          $ref: "#/components/schemas/PageableObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        content:
          items:
            $ref: "#/components/schemas/ProcessSensitiveVariables"
          type: array
        number:
          format: int32
          type: integer
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    ResponseDTOListProcessSensitiveVariables:
      example:
        data:
        - processDefinitionId: processDefinitionId
          createdAt: 2000-01-23T04:56:07.000+00:00
          variables:
          - name: name
            type: HIDE
            value: value
          - name: name
            type: HIDE
            value: value
          id: id
          enabled: true
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - processDefinitionId: processDefinitionId
          createdAt: 2000-01-23T04:56:07.000+00:00
          variables:
          - name: name
            type: HIDE
            value: value
          - name: name
            type: HIDE
            value: value
          id: id
          enabled: true
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          items:
            $ref: "#/components/schemas/ProcessSensitiveVariables"
          type: array
...
    ResponseDTOResourceDefinitionDTO:
      example:
        data:
          resourceMetadata:
            key: ""
          createdAt: 2000-01-23T04:56:07.000+00:00
          resourceXML: resourceXML
          version: 5
          definitionId: definitionId
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          $ref: "#/components/schemas/ResourceDefinitionDTO"
...
    ResponseDTOVoid:
      example:
        data: ""
        meta:
          code: 0
          message: message
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        data: {}
...